api/requirements.txt
SQLAlchemy==1.4.41
GeoAlchemy2==0.13.1
numpy==1.24.2
sqlmodel==0.0.8
uvicorn==0.20.0
psycopg2==2.9.5
python-dotenv==1.0.0